Addressing some duplication across generators
  - Moves identical methods to a base class
  - Changes all generators to inherit from base class
  - Cleans up some excess whitespace in templates
  - Fixes namespacing in plutus_generator.rb
