+++ lib/generators/plutus/base_generator.rb
@@ -1,0 +1,19 @@
+module Plutus
+  class BaseGenerator < Rails::Generators::Base
+    include Rails::Generators::Migration
+
+    def self.source_root
+      @source_root ||= File.join(File.dirname(__FILE__), 'templates')
+    end
+
+    # Implement the required interface for Rails::Generators::Migration.
+    # See http://apidock.com/rails/ActiveRecord/Generators/Base/next_migration_number/class
+    def self.next_migration_number(dirname)
+      if ActiveRecord::Base.timestamped_migrations
+        Time.now.utc.strftime("%Y%m%d%H%M%S")
+      else
+        "%.3d" % (current_migration_number(dirname) + 1)
+      end
+    end
+  end
+end
