+++ lib/generators/plutus/templates/migration.rb
@@ -23,7 +23,7 @@ class CreatePlutusTables < ActiveRecord::Migration
       t.references :account
       t.references :entry
       t.decimal :amount, :precision => 20, :scale => 10
-    end 
+    end
     add_index :plutus_amounts, :type
     add_index :plutus_amounts, [:account_id, :entry_id]
     add_index :plutus_amounts, [:entry_id, :account_id]
